refactor: rearrange folder structure of the app
