+++ frontend/src/modules/patient_analysis/components/index.ts
