+++ frontend/src/modules/patient_analysis/pages/PatientAnalysisPage.tsx
@@ -1,0 +1,12 @@
+import { AnalysisForm } from "@modules/patient_analysis/components";
+import { AnalysisFormProvider } from "@modules/patient_analysis/contexts";
+
+export default function PatientAnalysisPage() {
+  return (
+    <div>
+      <AnalysisFormProvider>
+        <AnalysisForm />
+      </AnalysisFormProvider>
+    </div>
+  );
+}
