--- frontend/src/pages/PatientAnalysisPage.tsx
@@ -1,9 +1,0 @@
-import { AnalysisForm } from "@/components/PatientAnalysis";
-
-export default function PatientAnalysisPage() {
-  return (
-    <div>
-      <AnalysisForm />
-    </div>
-  );
-}
