+++ frontend/src/routes/router.tsx
@@ -3,21 +3,12 @@ import {
   createBrowserRouter,
   createRoutesFromElements,
 } from "react-router-dom";
-import PatientAnalysisPage from "@/pages/PatientAnalysisPage";
-import RootLayout from "@/layout/RootLayout";
-import { AnalysisFormProvider } from "@/contexts/AnalysisFormContext";
-
+import PatientAnalysisPage from "@/modules/patient_analysis/pages/PatientAnalysisPage";
+import RootLayout from "@/layouts/RootLayout";
 export const router = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<RootLayout />}>
-      <Route
-        index
-        element={
-          <AnalysisFormProvider>
-            <PatientAnalysisPage />
-          </AnalysisFormProvider>
-        }
-      />
+      <Route index element={<PatientAnalysisPage />} />
     </Route>
   )
 );
