+++ frontend/src/shared/components/InputNumber.tsx
