+++ frontend/src/shared/types/utility.ts
